change account-manager to auth and moved in to pages
